change structure of application
