--- src/app/budget/[month]/MonthBudgetClient.tsx
@@ -1,23 +1,0 @@
-"use client";
-import { useRouter } from "next/navigation";
-import { PageHeader } from "./components/PageHeader";
-import { TransactionForm } from "./components/TransactionForm";
-import { TransactionTable } from "./components/TransactionTable";
-import { TransactionsSummary } from "./components/TransactionsSummary";
-
-interface Props {
-  month: string;
-}
-
-export default function MonthBudgetClient({ month }: Props) {
-  const router = useRouter();
-
-  return (
-    <main className="p-6">
-      <PageHeader month={month} />
-      <TransactionForm />
-      <TransactionsSummary />
-      <TransactionTable />
-    </main>
-  );
-}
