+++ src/app/layout.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from "next";
 import localFont from "next/font/local";
-import Header from "./components/Header/Header";
+import Header from "../components/Header/Header";
 
 import "./globals.css";
